Adding annotations @BeforeMethod and @AfterMethod to the base class
