+++ src/test/java/screenshot/Listeners.java
@@ -1,16 +1,13 @@
 package screenshot;
 
-import tests.BaseTest;
 import base.ExtentReporterNG;
 import com.aventstack.extentreports.ExtentReports;
 import com.aventstack.extentreports.ExtentTest;
 import com.aventstack.extentreports.Status;
-import org.openqa.selenium.WebDriver;
 import org.testng.ITestContext;
 import org.testng.ITestListener;
 import org.testng.ITestResult;
-
-import java.io.IOException;
+import tests.BaseTest;
 
 public class Listeners extends BaseTest implements ITestListener {
 
@@ -36,23 +33,24 @@ public class Listeners extends BaseTest implements ITestListener {
         extent.flush();
     }
 
-    @Override
-    public void onTestFailure(ITestResult result) {
-
-        WebDriver driver = null;
-        String testMethodName = result.getMethod().getMethodName();
-
-        extentTest.get().fail(result.getThrowable());
-
-        try {
-            driver = (WebDriver) result.getTestClass().getRealClass().getDeclaredField("driver").get(result.getInstance());
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-        try {
-            extentTest.get().addScreenCaptureFromPath(getScreenshot(testMethodName, driver), result.getMethod().getMethodName());
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
+    //TODO fix implementation of screenshots for extent report on failure
+//    @Override
+//    public void onTestFailure(ITestResult result) {
+//
+//        WebDriver driver = null;
+//        String testMethodName = result.getMethod().getMethodName();
+//
+//        extentTest.get().fail(result.getThrowable());
+//
+//        try {
+//            driver = (WebDriver) result.getTestClass().getRealClass().getDeclaredField("driver").get(result.getInstance());
+//        } catch (Exception e) {
+//            e.printStackTrace();
+//        }
+//        try {
+//            extentTest.get().addScreenCaptureFromPath(getScreenshot(testMethodName, driver), result.getMethod().getMethodName());
+//        } catch (IOException e) {
+//            e.printStackTrace();
+//        }
+//    }
 }
