+++ src/test/java/tests/BaseTest.java
@@ -2,6 +2,8 @@ package tests;
 
 import io.github.bonigarcia.wdm.WebDriverManager;
 import org.apache.commons.io.FileUtils;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.openqa.selenium.OutputType;
 import org.openqa.selenium.TakesScreenshot;
 import org.openqa.selenium.WebDriver;
@@ -10,6 +12,8 @@ import org.openqa.selenium.chrome.ChromeOptions;
 import org.openqa.selenium.edge.EdgeDriver;
 import org.openqa.selenium.firefox.FirefoxDriver;
 import org.openqa.selenium.safari.SafariDriver;
+import org.testng.annotations.AfterMethod;
+import org.testng.annotations.BeforeMethod;
 
 import java.io.File;
 import java.io.FileInputStream;
@@ -18,6 +22,7 @@ import java.util.Properties;
 import java.util.concurrent.TimeUnit;
 
 public class BaseTest {
+    public static Logger log = LogManager.getLogger(BaseTest.class.getName());
     public WebDriver driver;
     public Properties properties;
 
@@ -58,5 +63,17 @@ public class BaseTest {
         String destinationFile = System.getProperty("user.dir") + "/reports/" + testMethodName + ".png";
         FileUtils.copyFile(file, new File(destinationFile));
         return destinationFile;
+    }
+
+    @BeforeMethod
+    public void initializeTest() throws IOException {
+        driver = initializeDriver();
+        driver.get(properties.getProperty("url"));
+        log.info("Driver successfully initialized");
+    }
+
+    @AfterMethod
+    public void stopDriver() {
+        driver.quit();
     }
 }
