+++ src/test/java/tests/HomePageTest.java
@@ -2,13 +2,8 @@ package tests;
 
 import model.LandingPage;
 import model.LoginPage;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
 import org.testng.Assert;
-import org.testng.annotations.AfterTest;
-import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 
@@ -16,8 +11,6 @@ import java.io.IOException;
 
 
 public class HomePageTest extends BaseTest {
-    public static Logger log = LogManager.getLogger(BaseTest.class.getName());
-    public WebDriver driver;
 
     @DataProvider
     public Object[][] getData() {
@@ -28,19 +21,6 @@ public class HomePageTest extends BaseTest {
         data[1][0] = "tomsmith";
         data[1][1] = "SuperSecretPassword!";
         return data;
-    }
-
-    @BeforeMethod
-    public void initializeTest() throws IOException {
-
-        driver = initializeDriver();
-        driver.get(properties.getProperty("url"));
-        log.info("Driver successfully initialized");
-    }
-
-    @AfterTest
-    public void stopDriver() {
-        driver.quit();
     }
 
     @Test(dataProvider = "getData")
